+++ client/ui/Credits.jsx
@@ -32,11 +32,16 @@ export default class Credits extends Component {
                       <h5><b>Release information</b></h5>
 
                       <ul>
+                        <b>5 </b>
+                        <li>English Topics Search added</li>
+                      </ul>
+
+                      <ul>
                         <b>4 </b>
                         <li>Arabic Grammar added: Nouns, Proper Nouns, Verbs and Adjectives</li>
                         <li>Arabic Grammar also added for Noor Majidi script</li>
                         <li>Word for word translation added for English Corpus to Arabic</li>
-                        <li>Word for word translation added for Arabic to English Corpus</li>                        
+                        <li>Word for word translation added for Arabic to English Corpus</li>
                       </ul>
 
                       <ul>